move slot actions to new file
